HTTPHeadersLoad just swallow large/too many headers, instead of failing.
This is probably a good idea when parsing requests.
It's a VERY good idea when parsing responses. Content-Security-Policy headers can be huge, and sites like GitHub serve 22+ headers, the vast majority of which we don't care about.

To be fair, this isn't a perfect solution, since important headers can come at the end and get dropped.